+++ CHANGELOG.md
@@ -22,6 +22,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 - Snyk: Security upgrade com.amazonaws:aws-java-sdk-s3 from 1.12.378 to 1.12.386
 
+## [8.2.1]
+### Added
+### Added
+### Deprecated
+### Removed
+### Fixed
+- **PODAAC-5219**
+  - NumberUtils.createInteger(String:s) while s has leading zero, the NumberUtils will treat the string as Octal
+  - implemented a utility function to remove leading zero.
+### Security
+
 ## [8.2.0]
 ### Added
 ### Deprecated